Moved mock objects to anotehr file and finally got generate working.
